backend/requirements.txt
fastapi
uvicorn
python-dotenv
requests
msal
mysql-connector-python
bcrypt
pydantic[email]
PyJWT
google-auth
google-auth-oauthlib
google-auth-httplib2
